Syntax errors fixed :D

+++ multipath/yen.go
@@ -18,25 +18,39 @@ type metricPath struct {
 	metric float64
 }
 
-func (m []metricPath) Len() int {return len(m);}
-func (m []metricPath) Less(i, j int) bool {return m[i].metric < m[j].metric;}
-func (m []metricPath) Swap(i, j int) {
+type metricPathSlice []metricPath;
+func (m metricPathSlice) Len() int {return len(m);}
+func (m metricPathSlice) Less(i, j int) bool {return m[i].metric < m[j].metric;}
+func (m metricPathSlice) Swap(i, j int) {
 	temp := m[i];
 	m[i] = m[j];
 	m[j] = temp;
 }
 
+func pathsEqual(a, b []string) bool {
+	if len(a) != len(b) {
+		return false;
+	}
+	for idx, ia := range(a) {
+		if ia != b[idx] {
+			return false;
+		}
+	}
+	return true;
+}
 
-func Yen(grph graph.DefaultGraph, source, sink string, K int) [][]string {
+func Yen(grph graph.Graph, source, sink string, K int) [][]string {
 	var err error;
+	var firstPath []string;
 	A := make([][]string, 0, K);
-	A[0], _, err = graph.Dijkstra(grph, source, sink);
+	firstPath, _, err = graph.Dijkstra(grph, source, sink);
+	A = append(A, firstPath);
 	if err != nil {
 		return nil;
 	}
-	B := make([]metricPath);
-	removedNodes := make([]string);
-	removedEdges := make([]nodePair);
+	B := make(metricPathSlice, 0);
+	removedNodes := make([]string, 0);
+	removedEdges := make([]nodePair, 0);
 	for k := 1; k < K; k++ {
 		for i := 0; i < len(A[k - 1]); i++ {
 			spurNode := A[k - 1][i];
@@ -45,21 +59,21 @@ func Yen(grph graph.DefaultGraph, source, sink string, K int) [][]string {
 				if path == nil {
 					continue;
 				}
-				if rootPath == path[:i] {
-					weight, _ := grph.GetWeight(path[1], path[i+1]);
+				if i + 1 < len(path) && pathsEqual(rootPath, path[:i]) {
+					weight, _ := grph.GetWeight(path[i], path[i+1]);
 					grph.DeleteEdge(path[i], path[i+1])
 					removedEdges = append(removedEdges, nodePair{parent: path[i], child: path[i+1], weight: weight});
 				}
 			}
 			for _, node := range(rootPath) {
 				if node != spurNode {
-					neighbors, _ := graph.GetParents(node);
-					for _, neighbor := range(neighbors) {
+					neighbors, _ := grph.GetParents(node);
+					for neighbor, _ := range(neighbors) {
 						weight, _ := grph.GetWeight(neighbor, node);
 						removedEdges = append(removedEdges, nodePair{parent: neighbor, child: node, weight: weight});
 					}
-					neighbors, _ = graph.GetChildren(node);
-					for _, neighbor := range(neighbors) {
+					neighbors, _ = grph.GetChildren(node);
+					for neighbor, _ := range(neighbors) {
 						weight, _ := grph.GetWeight(node, neighbor);
 						removedEdges = append(removedEdges, nodePair{parent: node, child: neighbor, weight: weight});
 					}
@@ -85,7 +99,7 @@ func Yen(grph graph.DefaultGraph, source, sink string, K int) [][]string {
 			break;
 		}
 		sort.Sort(B);
-		A[k] = B[0];
+		A = append(A, B[0].path);
 		B = B[1:];
 	}
 	return A;
